+++ include/keyboard.h
@@ -11,7 +11,7 @@
 /**
  * 特殊キー
  */
-enum Key {
+enum key {
   UP        = 0x48,
   DOWN      = 0x50,
   BACKSPACE = 0x0E
@@ -19,7 +19,7 @@ enum Key {
 
 #include <lib/queue.h>
 
-struct KEYBUF {
+struct keybuffer {
   char data;
   unsigned char flag;
 } keybuf;