+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## ver 1.14.0
 - improve timer intr handler
-- rename queue
+- rename struct
 
 ## ver 1.13.0
 - support multiple timer